Implement Faculty, Room, and Building with tests

+++ src/main/java/edu/sandiego/comp305/sp24/schoolSim/model/Person.java
@@ -6,7 +6,7 @@ import edu.sandiego.comp305.sp24.schoolSim.Database;
 import java.sql.*;
 
 public abstract class Person {
-    private int id;
+    private long id;
     private String firstName;
     private String lastName;
     private Date birthdate;
@@ -18,7 +18,7 @@ public abstract class Person {
     private Department department;
 
     // Obtain a Person entry from the database
-    public Person(int id) {
+    public Person(long id) {
         String sql = "SELECT * FROM Person WHERE id=?";
         try {
             PreparedStatement preparedStatement = Database.getInstance().getDatabaseConnection().prepareStatement(sql);
@@ -45,7 +45,15 @@ public abstract class Person {
     }
 
     // Create a new Person object and insert it into the database
-    public Person(String firstName, String lastName, Date birthdate, String phoneNumber, String username, String organizationEmail, String secondaryEmail, boolean isActive, Department department) {
+    public Person(String firstName,
+                  String lastName,
+                  Date birthdate,
+                  String phoneNumber,
+                  String username,
+                  String organizationEmail,
+                  String secondaryEmail,
+                  boolean isActive,
+                  Department department) {
         this.firstName = firstName;
         this.lastName = lastName;
         this.birthdate = birthdate;
@@ -69,7 +77,7 @@ public abstract class Person {
             preparedStatement.setString(6, organizationEmail);
             preparedStatement.setString(7, secondaryEmail);
             preparedStatement.setBoolean(8, isActive);
-            preparedStatement.setInt(9, department.getId());
+            preparedStatement.setLong(9, department.getId());
 
             int affectedRows = preparedStatement.executeUpdate();
 
@@ -79,7 +87,7 @@ public abstract class Person {
 
             try (ResultSet generatedKeys = preparedStatement.getGeneratedKeys()) {
                 if (generatedKeys.next()) {
-                    this.id = (int) generatedKeys.getLong(1);
+                    this.id = generatedKeys.getLong(1);
                 } else {
                     throw new SQLException("Creating user failed, no ID obtained.");
                 }
@@ -89,7 +97,7 @@ public abstract class Person {
         }
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
