+++ src/test/java/edu/sandiego/comp305/sp24/schoolSim/model/StudentTest.java
@@ -95,7 +95,7 @@ class StudentTest {
         deleteWithId(student.getId(), "Student");
     }
 
-    void deleteWithId(int id, String table) {
+    void deleteWithId(long id, String table) {
         try {
             String sql = "DELETE FROM " + table + " WHERE `id` = ?";
